step3 - refactor view 는 package 를 분리
