+++ src/main/java/study/racing/view/ResultView.java
@@ -1,16 +1,16 @@
-package study.racing;
+package study.racing.view;
 
 import java.util.List;
 import java.util.stream.IntStream;
 
 public class ResultView {
 
-    public void printResult(int round, List<Integer> cars) {
-        for (Integer number : cars) {
+    public void printResult(List<Integer> carRecords) {
+        for (Integer number : carRecords) {
             IntStream.range(0, number)
                     .forEach(t -> System.out.print("-"));
             System.out.println();
         }
-        System.out.printf("End %d Round\n", round);
+        System.out.println("End Round");
     }
 }
